MySQL Conversion
The parent side of the application has been fully converted to use
MySQL, and all traces of SQLite have been removed. All previous
functionality is fully working with the MySQL database.

+++ Database/childcare_v4/childcare_v4_create.sql
@@ -44,7 +44,7 @@ CREATE TABLE ChildCareTransaction (
     Date date    NOT NULL ,
     CheckedIn time    NOT NULL ,
     CheckedOut time    NULL ,
-    TransactionTotal int    NULL ,
+    TransactionTotal float    NULL ,
     CONSTRAINT ChildCareTransaction_pk PRIMARY KEY (ChildCareTransaction_ID)
 );
